Add files via upload
the vmstat 1200 – monitors every 24 hours and puts the data into the vmstat1.data
grep “swap”- the swap should always be zero if its not then some process has consumed massive
memory. That will be monitored in this line
grep “r”- the running queue is constantly above process 1 it indicates the system is slow and
some process is waiting to be executed. That will be monitored here. Grep “cpu”- it indicates the cpu usage of the system. If the cpu usage is more it will be
monitored and will alert in this line.